+++ killstats/managers/killboard_manager.py
@@ -153,7 +153,7 @@ class KillmailBaseManager(models.Manager):
             region_id = killmail_body.get_or_create_region_id(
                 killmail_body.solar_system_id
             )
-            victim_ship = ItemType.objects.get(killmail_body.victim.ship_type_id)
+            victim_ship = ItemType.objects.get(id=killmail_body.victim.ship_type_id)
             victim = None
 
             if killmail_body.victim.character_id: